# Avoid allocating a new slice on every Append