@prashantv 2017-04-10
benchcmp:
```
benchmark                                                          old ns/op     new ns/op     delta
BenchmarkAppend/append_nil_to_nil/1-8                              6.07          6.24          +2.80%
BenchmarkAppend/append_nil_to_nil/2-8                              8.98          9.17          +2.12%
BenchmarkAppend/append_nil_to_nil/10-8                             37.7          38.0          +0.80%
BenchmarkAppend/append_single_error_to_nil/1-8                     6.11          6.18          +1.15%
BenchmarkAppend/append_single_error_to_nil/2-8                     78.0          71.4          -8.46%
BenchmarkAppend/append_single_error_to_nil/10-8                    1576          696           -55.84%
BenchmarkAppend/append_multiple_errors_to_nil/1-8                  6.10          6.28          +2.95%
BenchmarkAppend/append_multiple_errors_to_nil/2-8                  239           233           -2.51%
BenchmarkAppend/append_multiple_errors_to_nil/10-8                 3638          3570          -1.87%
BenchmarkAppend/append_nil_to_single_error/1-8                     6.76          6.25          -7.54%
BenchmarkAppend/append_nil_to_single_error/2-8                     9.72          9.17          -5.66%
BenchmarkAppend/append_nil_to_single_error/10-8                    39.5          39.1          -1.01%
BenchmarkAppend/append_single_error_to_single_error/1-8            75.9          68.0          -10.41%
BenchmarkAppend/append_single_error_to_single_error/2-8            239           182           -23.85%
BenchmarkAppend/append_single_error_to_single_error/10-8           1786          741           -58.51%
BenchmarkAppend/append_multiple_errors_to_single_error/1-8         201           192           -4.48%
BenchmarkAppend/append_multiple_errors_to_single_error/2-8         447           430           -3.80%
BenchmarkAppend/append_multiple_errors_to_single_error/10-8        3969          3847          -3.07%
BenchmarkAppend/append_nil_to_multiple_errors/1-8                  6.40          6.51          +1.72%
BenchmarkAppend/append_nil_to_multiple_errors/2-8                  9.29          9.36          +0.75%
BenchmarkAppend/append_nil_to_multiple_errors/10-8                 39.2          39.8          +1.53%
BenchmarkAppend/append_single_error_to_multiple_errors/1-8         204           198           -2.94%
BenchmarkAppend/append_single_error_to_multiple_errors/2-8         402           378           -5.97%
BenchmarkAppend/append_single_error_to_multiple_errors/10-8        2198          718           -67.33%
BenchmarkAppend/append_multiple_errors_to_multiple_errors/1-8      238           232           -2.52%
BenchmarkAppend/append_multiple_errors_to_multiple_errors/2-8      522           508           -2.68%
BenchmarkAppend/append_multiple_errors_to_multiple_errors/10-8     4180          4129          -1.22%
```

And number of allocs:
```
benchmark                                                          old allocs     new allocs     delta
BenchmarkAppend/append_nil_to_nil/1-8                              0              0              +0.00%
BenchmarkAppend/append_nil_to_nil/2-8                              0              0              +0.00%
BenchmarkAppend/append_nil_to_nil/10-8                             0              0              +0.00%
BenchmarkAppend/append_single_error_to_nil/1-8                     0              0              +0.00%
BenchmarkAppend/append_single_error_to_nil/2-8                     2              2              +0.00%
BenchmarkAppend/append_single_error_to_nil/10-8                    26             13             -50.00%
BenchmarkAppend/append_multiple_errors_to_nil/1-8                  0              0              +0.00%
BenchmarkAppend/append_multiple_errors_to_nil/2-8                  3              3              +0.00%
BenchmarkAppend/append_multiple_errors_to_nil/10-8                 27             27             +0.00%
BenchmarkAppend/append_nil_to_single_error/1-8                     0              0              +0.00%
BenchmarkAppend/append_nil_to_single_error/2-8                     0              0              +0.00%
BenchmarkAppend/append_nil_to_single_error/10-8                    0              0              +0.00%
BenchmarkAppend/append_single_error_to_single_error/1-8            2              2              +0.00%
BenchmarkAppend/append_single_error_to_single_error/2-8            5              4              -20.00%
BenchmarkAppend/append_single_error_to_single_error/10-8           29             14             -51.72%
BenchmarkAppend/append_multiple_errors_to_single_error/1-8         3              3              +0.00%
BenchmarkAppend/append_multiple_errors_to_single_error/2-8         6              6              +0.00%
BenchmarkAppend/append_multiple_errors_to_single_error/10-8        30             30             +0.00%
BenchmarkAppend/append_nil_to_multiple_errors/1-8                  0              0              +0.00%
BenchmarkAppend/append_nil_to_multiple_errors/2-8                  0              0              +0.00%
BenchmarkAppend/append_nil_to_multiple_errors/10-8                 0              0              +0.00%
BenchmarkAppend/append_single_error_to_multiple_errors/1-8         3              3              +0.00%
BenchmarkAppend/append_single_error_to_multiple_errors/2-8         6              5              -16.67%
BenchmarkAppend/append_single_error_to_multiple_errors/10-8        30             13             -56.67%
BenchmarkAppend/append_multiple_errors_to_multiple_errors/1-8      3              3              +0.00%
BenchmarkAppend/append_multiple_errors_to_multiple_errors/2-8      6              6              +0.00%
BenchmarkAppend/append_multiple_errors_to_multiple_errors/10-8     30             30             +0.00%
```